+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Internal changes:
   `python-api` to discover what microcontroller the keyboard is using.
 * `firmware` rework flash writing commands so that is possible to update only a
   portion of the layout.
+* `firmware` layout updates can now be performed without the USB device
+  needing to be reset
 * `firmware-xmega` now utilizes the IRQ pin in receive mode
 * `firmware-xmega` add arbitrary pin mapping to matrix scanning algorithm. It
   should now be possible to use any pins as row/columns.